refactor: moved StacAssetExplorer to stac folder
