+++ sites/geohub/src/components/stac/StacAssetExplorer.svelte
